JAVA CONSOL OOP FIX
